added service test

+++ Backend/src/main/java/com/cabbooking/model/AbstractUser.java
@@ -72,8 +72,8 @@ public abstract class AbstractUser {
         return id;
     }
 
-    public void setId(Integer id) {
-        this.id = id;
+    public void setId(long id) {
+        this.id = (int) id;
     }
 
     public String getUsername() {
